requirements.txt
openai-whisper
fastapi[standard]
python-vlc
git+https://github.com/dhanushnatra/piper-nex.git
